Refactor hydration and  UI update
