+++ Stanford360/Resources/Localizable.xcstrings
@@ -20,9 +20,6 @@
     "%lld" : {
 
     },
-    "%lld Day Streak" : {
-
-    },
     "%lld Meals Logged" : {
 
     },
@@ -719,9 +716,6 @@
 
     },
     "Streak" : {
-
-    },
-    "Streak icon" : {
 
     },
     "Swift Package Manager" : {
